change to float

+++ fairmotion/models/SpatioTemporalTransformer.py
@@ -76,7 +76,7 @@ class SpatialTemporalEncoderLayer(nn.Module):
         tx = torch.reshape(tx, (T, B*S, E)) # T x (B*S) x E
         # tx: input to temporal multihead (T, B*S, E)
         # t_mask : only need to mask for temporal attention not spatial
-        t_mask = self._generate_square_subsequent_mask(T).to(device=x.device,)
+        t_mask = self._generate_square_subsequent_mask(T).to(device=x.device, dtype=torch.float64)
         tm, _ = self.TemporalMultiheadAttention(tx, tx, tx, attn_mask= t_mask) # T x (B*S) x E
         tm = self.dropout_1(tm)
         tm = self.norm_1(tm + tx)
